Initial things for first Spring
Implemented the basic interpreter that can read from the test.pseudo file and output it into the outfile.py file. The interpreter reads each line runs the keyword handler in the interlib folder. For now, we only have Create so the create.py file handles all of the Create keywords.