Made changes to the internal typing of variable
Made it so that in the all_variables{} dictionary each variable's data_type is more specific. A variable can now be a "number" or a "string". This makes it easier for the ones who are doing the arithmetic functions to check whether or not a variable is of the correct type to be added.