Updated Display and created a utility file
The "Display" keyword now supports integers and floats. All "Display" features should now be done.
A utility file that contains useful helper functions. For now it contains a function called inter_data_type(). Takes in a single data string and returns a string determining what type of data it is. Will definitely be useful for implementing the arithmetic functions