Added a filtration thing for each line
Filters out newline characters and empty strings when there are spaces after the last word of the line. This should make parsing each line much cleaner and hopefully faster. Matches line positions for both Pseudo and Python code. eg: If there are 10 lines that are empty before this line "Create a variable named x with a value 1" and it is on line 20, then "x = 1" would also be on line 20 instead of 10 lines before. Kind of hard to describe it, but add empty lines in between your Pseudo code to understand what I'm talking about.