Final commit, hopefully
Sanity commit for the interpreter and testfiles. We'll be using the output.py file to display the Python code to the python section of the GUI and the output.txt file to the output section of the GUI. The test.pseudo file contains test cases to check if any of the keywords are erroring out, but it seems like everything is working.

Got list and tables to hold list and tables in a hacky solution, but it'll work for the "Display" keyword.