Check empty pseudo files and modify outputs
Made the interpreter handle errors for empty files and fixed how the outputs should display to the console of the GUI. Will probably need to update the code so that it is more optimized. Can be easily changed to a function by switching out the comments near the main function declaration.