Added more variables to the interpret_state
I figured that it would be impossible to check if the current line is in a function, if, loop, or while keywords based on how we are iterating each line one by one. The interpreter should now have the ability to jump ahead of code when we are parsing. Meaning that we can have our own parsing loop in the keyword handlers and then assign the new program counter to interpret_state["line_numb"] so that we can continue to parse normally.