Update interpreter.py
Added more stuff to the interpreter_state dictionary so that the Import keyword can work. Later on when functions are implemented, I'll make it so that the import keyword will import those specific functions. For now it will do "import <file_name>", but later on it will do "from <file_name> import <function_name>"