More checks for comments and pycode
Made comments and pycode more malleable so that you can have comments that start with a # in the word itself and also used a special character for pycode.
The % character will denote a line to be a pycode as well. For instance: "%x = 1" is the same as "Pycode x = 1".
Added a except message when a user has invalid usages of the Pycode/%. Unfortunate that it doesn't detect which line the error occurred.