Added a thing to detect closing of the GUI
Added a feature where when a user closes the GUI, the whole program closes.
Needed because the "Live Update" feature would still have a reference to the loaded file and if the GUI closes then the input text would be all cleared. However, this doesn't mean that the "Live Update" stops terminating and the opened file will be overwritten, clearing the file of its contents. This protocol stops this.