Fixed Settings option
The Settings menu could not be open because of how the setting was trying to reference the root of the GUI since the splitting of the files. Now this should reference the master of the GUI (which is basically the same thing as the root) and added importing thing to make it work.