Update add.py
There was an unexpected bug where writing "Add 1 and 2, store into" resulted the python code to be " into = 1 + 2", creating a variable named "into" when it should actually be an error. Fixed it, but I had to refactor most the code so that specific cases where article words would error out if they're not in the pseudo code.