Fixed the float problem
Apparently I had the period character (.) to be included on whether or not the value list is a number. So every time that a float number string is passed into the num_parse() function it'll assume that the float number is not a number at all. I also undid my indent at the end of the for loop that determines if the value_list[0] is a number or not. Not sure why I indented it in the first place.