Added debugging feature
There is a toggle in Help menu that if true will display in the console python code errors.
Currently exceptions thrown in the application.py code will be shown, and callback exceptions.
Whether it is toggled or not those exceptions will also be appended to an error log. This can be helpful for developers as well.