Found file except when attempt to run empty file
Found an exception where if you try to execute an empty file, you would get an error exception if you did not save it because the interpreter requires a file pointer object. Made that an if conditional.
Also removed the debugging messages from the Help Document's try-except code since it'll print tons of useless exception messages.